Authorization

+++ app/Http/Controllers/PostController.php
@@ -6,10 +6,22 @@ use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
+    /**
+     * Get the middleware that should be assigned to the controller.
+     */
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth', except: ['index', 'show']),
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -61,6 +73,9 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        // Authorizing the action
+        Gate::authorize('modify', $post);
+
         return view('posts.edit', ['post' => $post ]);
     }
 
@@ -69,6 +84,9 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        // Authorizing the action
+        Gate::authorize('modify', $post);
+
         // Validate
         $fields = $request->validate([
             'title' => ['required', 'max:255'],
@@ -87,6 +105,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        // Authorizing the action
+        Gate::authorize('modify', $post);
+        
         // Delete the post
         $post->delete();
